# server/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_costrict_costrict-web_server_internal_models.CapabilityArtifact:
    properties:
      artifactVersion:
        type: string
      checksumSha256:
        type: string
      createdAt:
        type: string
      downloadCount:
        type: integer
      fileSize:
        type: integer
      filename:
        type: string
      id:
        type: string
      isLatest:
        type: boolean
      itemId:
        type: string
      mimeType:
        type: string
      sourceType:
        type: string
      storageBackend:
        type: string
      storageKey:
        type: string
      uploadedBy:
        type: string
    type: object
  github_com_costrict_costrict-web_server_internal_models.CapabilityAsset:
    properties:
      contentSha:
        type: string
      createdAt:
        type: string
      fileSize:
        type: integer
      id:
        type: string
      itemId:
        type: string
      mimeType:
        type: string
      relPath:
        type: string
      storageBackend:
        type: string
      storageKey:
        type: string
      textContent:
        type: string
      updatedAt:
        type: string
    type: object
...
  github_com_costrict_costrict-web_server_internal_models.CapabilityRegistry:
    properties:
      createdAt:
        type: string
      description:
        type: string
      externalBranch:
        type: string
      externalUrl:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityItem'
        type: array
      lastSyncLog:
        $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.SyncLog'
      lastSyncLogId:
        type: string
      lastSyncSha:
        type: string
      lastSyncedAt:
        type: string
      name:
        type: string
      ownerId:
        type: string
      repoId:
        type: string
      sourceType:
        type: string
      syncConfig:
        type: object
      syncEnabled:
        type: boolean
      syncInterval:
        type: integer
      syncStatus:
        description: idle | syncing | error | paused
        type: string
      updatedAt:
        type: string
    type: object
  github_com_costrict_costrict-web_server_internal_models.CapabilityVersion:
    properties:
      assets:
        items:
          $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityVersionAsset'
        type: array
      category:
        type: string
      commitMsg:
        type: string
      content:
        type: string
      contentMd5:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      id:
        type: string
      itemId:
        type: string
      metadata:
        type: object
      name:
        type: string
      revision:
        type: integer
      sourcePath:
        type: string
      version:
        type: string
    type: object
...
  github_com_costrict_costrict-web_server_internal_models.ItemTagDict:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: string
      slug:
        type: string
      tagClass:
        description: system | custom
        type: string
    type: object
...
  internal_handlers.MyItem:
    properties:
      artifacts:
        items:
          $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityArtifact'
        type: array
      assets:
        items:
          $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityAsset'
        type: array
      category:
        type: string
      content:
        type: string
      contentMd5:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      currentRevision:
        type: integer
      description:
        type: string
      embeddingUpdatedAt:
        type: string
      evaluation:
        type: object
      experienceScore:
        type: number
      favoriteCount:
        type: integer
      health:
        type: object
      id:
        type: string
      installCount:
        type: integer
      itemType:
        type: string
      lastScanId:
        type: string
      metadata:
        type: object
      name:
        type: string
      previewCount:
        type: integer
      registry:
        $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityRegistry'
      registryId:
        type: string
      repoId:
        type: string
      repoName:
        type: string
      repoVisibility:
        type: string
      securityStatus:
        type: string
      slug:
        type: string
      source:
        description: 导入来源，如 anthropic/claude-code, superpower, github
        type: string
      sourcePath:
        type: string
      sourceSha:
        type: string
      sourceType:
        description: direct | archive
        type: string
      status:
        type: string
      tags:
        description: Tags loaded separately (not persisted by GORM)
        items:
          $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.ItemTagDict'
        type: array
      updatedAt:
        type: string
      updatedBy:
        type: string
      version:
        type: string
      versions:
        items:
          $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityVersion'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    name: CoStrict Team
    url: https://costrict.ai
  description: AI Agent Platform API - Skill marketplace, repository and registry
    management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Costrict Web API
  version: "1.0"
paths:
...
  /items/my:
    get:
      description: Get all skill items owned by the current authenticated user
      parameters:
      - description: Filter by item type
        in: query
        name: type
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Search by name or description
        in: query
        name: search
        type: string
      - description: Filter by category (legacy single value)
        in: query
        name: category
        type: string
      - description: Filter by categories (comma-separated)
        in: query
        name: categories
        type: string
      - description: Filter by security statuses (comma-separated)
        in: query
        name: securityStatuses
        type: string
      - description: Filter by sources (comma-separated, OR logic)
        in: query
        name: source
        type: string
      - description: Filter by registry ID
        in: query
        name: registryId
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20, max: 100)'
        in: query
        name: pageSize
        type: integer
      - description: Sort by updatedAt, createdAt, previewCount, installCount, favoriteCount,
          or experienceScore
        in: query
        name: sortBy
        type: string
      - description: 'Sort order: asc or desc (default: desc)'
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              hasMore:
                type: boolean
              items:
                items:
                  $ref: '#/definitions/internal_handlers.MyItem'
                type: array
              page:
                type: integer
              pageSize:
                type: integer
              total:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      summary: List my items
      tags:
      - items
...
  /registries:
    get:
      description: 'Get registries visible to the current user. Visibility is determined
        by the parent repository''s visibility field: public repos'' registries are
        visible to everyone, private repos'' registries are visible to repo members,
        and registries without a repo are visible to their owner.'
      parameters:
      - description: Filter by repository ID
        in: query
        name: repoId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              registries:
                items:
                  $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityRegistry'
                type: array
            type: object
      summary: List registries
      tags:
      - registries
    post:
      consumes:
      - application/json
      description: Create a new skill registry. Visibility is inherited from the parent
        repository.
      parameters:
      - description: Registry data
        in: body
        name: body
        required: true
        schema:
          properties:
            description:
              type: string
            externalBranch:
              type: string
            externalUrl:
              type: string
            name:
              type: string
            ownerId:
              type: string
            repoId:
              type: string
            sourceType:
              type: string
            syncEnabled:
              type: boolean
            syncInterval:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityRegistry'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Create registry
      tags:
      - registries
  /registries/{id}:
    delete:
      description: Delete registry by ID
      parameters:
      - description: Registry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Delete registry
      tags:
      - registries
    get:
      description: Get registry by ID with its items
      parameters:
      - description: Registry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityRegistry'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get registry
      tags:
      - registries
    put:
      consumes:
      - application/json
      description: 'Update registry by ID. Note: visibility is controlled at the repository
        level, not the registry level.'
      parameters:
      - description: Registry ID
        in: path
        name: id
        required: true
        type: string
      - description: Registry data
        in: body
        name: body
        schema:
          properties:
            description:
              type: string
            externalBranch:
              type: string
            externalUrl:
              type: string
            name:
              type: string
            sourceType:
              type: string
            syncEnabled:
              type: boolean
            syncInterval:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityRegistry'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Update registry
      tags:
      - registries
...
  /registries/{id}/transfer:
    put:
      consumes:
      - application/json
      description: Transfer a registry's ownership to a different repository. Caller
        must be the registry owner_id or an admin/owner of the current repo, and at
        least a member of the target repo. Sync registries cannot be transferred while
        syncing.
      parameters:
      - description: Registry ID
        in: path
        name: id
        required: true
        type: string
      - description: Target repository ID
        in: body
        name: body
        required: true
        schema:
          properties:
            targetRepoId:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityRegistry'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Transfer registry to another repository
      tags:
      - registries
  /registries/my:
    get:
      description: Get all registries owned by a user
      parameters:
      - description: Owner user ID
        in: query
        name: ownerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              registries:
                items:
                  $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityRegistry'
                type: array
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      summary: List my registries
      tags:
      - registries
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"